Update CIS129_TyHuetter_Lab05
The code was scrapped and started over. I realized that I was doing the code in the wrong order. Many of the print statements needed to be in the 2nd while loop inside of the bigger while loop instead of vice versa. I also cleaned up the code and added more comments